change from meter to centimeter and modify code for new table.h

+++ source/sensor/lidar.h
@@ -28,7 +28,7 @@
 
 typedef struct Lidar {
 
-    double value;
+    uint16_t value;
 
     Timer* timer;
